add deploy subcommand
